+++ src/utils/const.js
@@ -1,6 +1,28 @@
+export const FILM_CARD_COUNT = 5;
+
 export const FilterType = {
   ALL: `All`,
   WATCHLIST: `watchlist`,
   HISTORY: `history`,
   FAVORITES: `favorites`
+};
+
+export const SortType = {
+  DEFAULT: `default`,
+  YEAR: `year`,
+  RATING: `rating`,
+  COMMENTS: `comments`,
+};
+
+export const UserRating = {
+  NONE: ``,
+  NOVICE: `Novice`,
+  FAN: `Fan`,
+  MOVIE_BUFF: `Movie Buff`,
+};
+
+export const Containers = {
+  MAIN: `Main`,
+  TOP: `Top`,
+  COMMENTED: `commented`,
 };
