+++ src/utils/render.js
@@ -6,6 +6,7 @@ export const RenderPosition = {
 };
 
 export const render = (container, child, place) => {
+
   if (container instanceof Abstract) {
     container = container.getElement();
   }
